GENAI/requirements.txt
python-dotenv
streamlit
openai
anthropic
google-generativeai
cohere